Moved submission templates to relevant directories
