--- api/src/middlewares/services/CaseService/templates/index.ts
@@ -1,4 +1,0 @@
-import marriageSubmission from "./marriageSubmission";
-import certifyCopySubmission from "./certifyCopySubmission";
-
-export { marriageSubmission, certifyCopySubmission };
